Update readme files, simply the app link workflow which can be used by iOS integration in the future.
