+++ native-apps/unity_ii_applink/unity_project/.gitignore
@@ -24,7 +24,6 @@
 
 # Gradle cache directory
 .gradle/
-gradle/
 
 # Autogenerated VS/MD/Consulo solution and project files
 ExportedObj/
