--- native-apps/unity_ii_applink/unity_project/Assets/Android.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2655c98839f2a3d45832406122f52a08
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
