--- native-apps/unity_ii_applink/unity_project/Assets/Android/Plugins.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9bc52b823e4726e4c992caabcb215c49
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
