--- native-apps/unity_ii_applink/unity_project/Assets/Android/Plugins/com.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c8fae1178b05d304f9b65266d78c7861
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
