+++ native-apps/unity_ii_applink/unity_project/README.md
@@ -3,9 +3,9 @@ This sample demonstrates one way to integrate Identity Integration with Unity on
 
 ## Overview
 
-In this example, you can learn how to communicate with the IC from C#. As we described in [Internet Identity Integration](../unity_android_deeplink/README.md#workflow), the game mainly focuses on:
+In this example, you can learn how to communicate with the IC from C#. As we described in [Internet Identity Integration](../README.md#workflow), the game mainly focuses on:
 
-1. Provide an Android Java plugin which can be a bridge between C# and the browser.
+1. Provide a plugin to handle the communication between C# and the browser.
 2. Register the App Links in the Android manifest file.
 3. Generate the `Ed25519KeyIdentity` in C# and pass the public key to the Web Brower.
 4. Receive the `DelegationIdentity` from the Web Brower and pass it from Java to C#.
@@ -15,14 +15,11 @@ In this example, you can learn how to communicate with the IC from C#. As we des
 
 There're several important files in this example.
 
-- [ICGameKitPlugin.java](./Assets/Android/Plugins/com/icgamekit/plugin/ICGameKitPlugin.java)  
-  This is a Unity Android native plugin written in java, provided with functions including opening browser, sending messages back to Unity.
-
 - [AndroidPostBuildProcessor.cs](./Assets/Editor/AndroidPostBuildProcessor.cs)  
   This is a Unity post-build processor for Android platform, which inserts the App Links URL scheme into the AndroidManifest.xml.
 
-- [PluginProxy.cs](./Assets/Scripts/PluginProxy.cs)  
-  PluginProxy works as a bridge between Unity C# and Android Java, which calls the functions inside [ICGameKitPlugin.java](./Assets/Android/Plugins/com/icgamekit/plugin/ICGameKitPlugin.java).
+- [DeepLinkPlugin.cs](./Assets/Scripts/DeepLinkPlugin.cs)  
+  It's a plugin which provides functions to open the web browser, handle the DeepLink activation, and parse the delegation from Json string.
 
 - [GreetingClient.cs](./Assets/Scripts/GreetingClient.cs)  
   This provides the C# APIs for calling the Greeting backend canister in [ii_integration_page](../ii_integration_page/README.md). It can be generated by running the `ClientGenerator` provided by [ICP.NET](https://github.com/BoomDAO/ICP.NET).
